+++ package.lisp
@@ -1,5 +1,13 @@
 (defpackage #:com.aragaer.pa-brain
   (:use #:common-lisp)
-  (:export :react)
-  (:export :handle-message)
-  (:export :*active-user*))
+  (:import-from #:alexandria
+                :compose
+                :if-let
+                :lastcar
+                :switch
+                :when-let)
+  (:export :react
+           :handle-message
+           :set-user
+           :*pa2human*
+           :*human2pa*))