+++ test.lisp
@@ -14,6 +14,7 @@
 
 (if (and
      (prove:run #P"tests/active_user.lisp" :reporter :list)
+     (prove:run #P"tests/commands.lisp" :reporter :list)
      (prove:run #P"tests/dont_understand.lisp" :reporter :list)
      (prove:run #P"tests/good_morning.lisp" :reporter :list)
      (prove:run #P"tests/handle_message.lisp" :reporter :list)