+++ tests/commands.lisp
@@ -2,33 +2,31 @@
 
 (load #P"brain.lisp")
 
-(defpackage good-morning-test
+(defpackage command-test
   (:use :cl
         :prove
         :com.aragaer.pa-brain))
 
-(in-package #:good-morning-test)
+(in-package #:command-test)
 (plan nil)
 
 (let* ((from '((:user . "user") (:channel . "channel")))
-       (side-channel '((:user . "user") (:channel . "incoming")))
        (to '((:user . "niege") (:channel . "brain")))
        (presence `((:event . "presence") (:from . ,from) (:to . ,to)))
-       (event `((:event . "new-day") (:to . ,to) (:from . ,from)))
+       (event `((:event . "new-day") (:to . ,to)))
        (result nil)
        (msg nil))
 
-  (subtest "with active channel"
+  (subtest "say with active channel"
 	   (set-user "user")
-	   (setf result (handle-message presence))
-	   (is result nil)
-	   (setf result (handle-message event))
+	   (handle-message presence)
+	   (setf result (handle-message '((:command . "say") (:text . "hello"))))
 	   (setf msg (car result))
-	   (is (assoc :message msg) (cons :message "good morning"))
+	   (is (assoc :message msg) (cons :message "hello"))
 	   (is (assoc :from msg) (cons :from to))
 	   (is (assoc :to msg) (cons :to from)))
 
-  (subtest "without active channel"
+  (when nil
 	   (set-user "user")
 	   (setf result (handle-message event))
 	   (is result nil)